back/requirements.txt
aiosqlite==0.22.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
APScheduler==3.11.2
asyncpg==0.31.0
asyncssh==2.22.0
bcrypt==5.0.0
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
et_xmlfile==2.0.0
fastapi==0.126.0
greenlet==3.3.0
h11==0.16.0
idna==3.11
invoke==2.2.1
numpy==2.4.0
openpyxl==3.1.5
pandas==2.3.3
paramiko==4.0.0
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.5
PyJWT==2.9.0
pydantic_core==2.41.5
PyNaCl==1.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-nmap==0.7.1
pytz==2025.2
reportlab==4.4.6
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uvicorn==0.38.0
websockets==15.0.1
